Your voice will make a difference.
Are you one of the organizations that are being exposed as all talk? just in it for a buck? Act now and avoid this exposure.

We have a project ending conflict in hundreds of countries. Our small local groups share ethical ad revenue with local people joining their local online group.

Once the small local group has enough people (about 50) they call a committee meeting get together to discuss the kiosk project.

Even if your not interested in theproject you should join to avoid your name appearing on The LIST Some say it's is a can't walk the talk LIST, We think ignoring mother nature in her hour of need is criminal.

Also this planet is about to become a nasty place to live if we don't stop ignoring the reality of massive Arctic methane release soon.

Solutions are available but the wrong people are governing the world. So we allow free membership and you can leave at anytime.

Mother Nature needs all of us caring people, she can be heard screaming in the face of the rape that's going on in front of our nose's, RiseUp THINK THINK THINK
--- --- --- --- --- --- --- --- --- --- --- --- --- --
http://api.ning.com/files/Sg8I27-Rz4wdufkdYZc6pzJZiA-cHao*yvgX22Sgn2E6Ak9HsojnQi4i1gRIiPUdM7awOKQKF3XzO-ermzxjc4zS1QnNz8rn/worldEconomicfourm40.jpg